Remove Ember route `refresh` event and manually add a new model to controller
Refresh was trying to pull models from the server that were deleted when
on a route that was using them, even when a transition to a parent route
was placed above the refresh call (probably has to do with Ember internal
transition promises). `refresh` necessity also appears inconsistent...
groups controller shows new model without refresh but items controller doesn't.